PARQUET-2391: Remove unnecessary unboxing (#1214)

+++ parquet-column/src/main/java/org/apache/parquet/schema/GroupType.java
@@ -175,7 +175,7 @@ public class GroupType extends Type {
     if (i == null) {
       throw new InvalidRecordException(name + " not found in " + this);
     }
-    return i.intValue();
+    return i;
   }
 
   /**
